feat(pubsub): subscriber write callback (fast-path)
Provide an optional write callback for subscriber in fast-path
configuration:
If the callback is set, the application handles the write operation.

This can be useful when a special "write" function (e.g. with signaling)
is needed to write the variable in the underlying engineering system.
Without the write callback the sdk would first memcpy
the value to a buffer and afterwards the application needs to
retrieve the value and use special "write" function.
If the write callback is set, then the application can directly write
the variable value with it's preferred implementation.

Note: the examples only work if raw data encoding is not used
in publish_rt example.