test(pubsub): subscriber write callback (fast-path)
For the subscriber write callback test we are using the test-plugin
implementation to properly simulate server pubsub callback
processing.

The test-plugin implements a fake clock. The current time
(UA_DateTime_now) always returns the same value. Therefore the
PublishedDataSet configuration version check does not work.
UA_PubSubConfigurationVersionTimeDifference() always
returns the same value.
Therefore we have to configured the PublishedDataSet before the
DataSetWriter, otherwise the DataSet information is not used, because
the DataSetMetaData version has not changed.