Fix segfault at checkReaderIdentifier
The function checkReaderIdentifier needs the
GroupHeader, WriterGroupId and DataSetWriterId in the received message
to find the correct reader.

The check at the beginning shall verify that all of these parameters
are available.
But the check was wrong and therefore a segfault happened later, if
one of the parameters was missing.